% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{detect_capabilities}
\alias{detect_capabilities}
\title{Detect which optional OMOP fields are available in the data source}
\usage{
detect_capabilities(connector)
}
\arguments{
\item{connector}{A \code{trajectory_connector} object.}
}
\value{
A modified \code{connector} with \verb{$capabilities} populated.
}
\description{
Detect which optional OMOP fields are available in the data source
}
