% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{disconnect_connector}
\alias{disconnect_connector}
\title{Close a persistent omop_connector database connection}
\usage{
disconnect_connector(connector)
}
\arguments{
\item{connector}{An \code{omop_connector} object.}
}
\value{
\code{connector} invisibly, with \verb{$conn} set to \code{NULL}.
}
\description{
Close a persistent omop_connector database connection
}
