% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_concepts.R
\docType{data}
\name{MYOSITIS_DRUG_CONCEPTS}
\alias{MYOSITIS_DRUG_CONCEPTS}
\title{OMOP concept IDs for myositis-relevant medications}
\format{
Named list of integer vectors.
}
\usage{
MYOSITIS_DRUG_CONCEPTS
}
\description{
A named list mapping drug family names to integer vectors of OMOP
drug_concept_id values (ingredient level).
}
\keyword{datasets}
